+++ controllers/likeController.js
@@ -20,12 +20,24 @@ const getLikesByUser = async (id) => {
     return data;
 }
 
-const createLike = async (data) => {
+const createLike = async (data, tokenUserId) => {
+    // cant create a like if the user has already liked
+    const likeData = await Like.findOne({where: {userId: data.userId, eventId: data.eventId}});
+    if (likeData) {
+        return 409;
+    }
+    if (Number(data.userId) !== tokenUserId) {
+        return 401;
+    }
     const like = await Like.create(data);
     return like;
 }
 
-const deleteLike = async (id) => {
+const deleteLike = async (id, tokenUserId) => {
+    const data = await Like.findOne({where: {id: id}});
+    if (Number(data.userId) !== tokenUserId) {
+        return 401;
+    }
     const like = await Like.destroy({where: {id: id}});
     return like;
 }
